distributor/storage: fix wrong nil error when writing fails on all peers.
When writing a block fails an all peers a nil error is returned (since
err is nil in that scope).

This cause the volume to fail opening again since BlockFile.Sync
thinks that f.File.SyncINode() wrote the inode block to at least one
store and so it'll write the inode metadata. On next volume opening the
inode block cannot be retrieved leaving the volume inaccessible.

This patch fixes distributor/storage.WriteBlock to return an error if
writing the block fails on all the peers.