server: create new lease if expired
If a server cannot communicate with etcd for a time greater then the lease
timeout (30s) its lease will expire and the server won't reregister
requiring it to be restarted.

This patch fixes this making the server, inside the heartbeat goroutine,
renew the current lease or create a new lease if expired and use this
one to register itself.