ring: Append Ketama nodes, not panic on small rings
The Ketama hash may not assign points to servers that are particularly
small compared to the total storage. This makes some sense -- they can't
cover much. However, this causes errors, and we need to return a full
permutation, so simply append the small ones as "overflow". Fixes #246.

Rings can also declare their replication factor as being higher than the
number of nodes they contain. This causes panics. Raise a warning when
such a ring is detected, but continue operation. Fixes #219.