Move actions from torusblk to torusctl. Fixes #206
In the past, it was separated by whether it was a volume action vs. a
generic metadata action. This is bad UX; most people want `torusctl` to
be simply the one tool they need.

This change turns `torusblk` into the client/mount-tool only for block
devices, true to its name. `torusd` is still the storage daemon, and
`torusctl` (analogous to `kubectl`) is your one-stop shop for working
with torus.