storage/block_device: initial implementation
This commit adds support for using a block device as the backing storage for
torusd.

Two utilities have been added, mkfs.torus and fsck.torus, to format a block
device for use by torus, and for checking the consistency of a formatted block
device, respectively.

There is a metadata section of size 4k bytes written to the beginning of the
disk, and the rest of the disk is used to store data.

Each location on the disk in which data can be stored has two things: block
headers of size 4k bytes, and the actual data, whose size is determined by the
torus cluster's block size.

When given a block ref, the location of the data is determined by feeding 24
bits comprised of the block ref's volume, inode, and index through sha256,
converting the last 8 bytes into a uint64, and modding that by the number of
slots available to store data in. By going to this location and reading the
block headers, one can then determine if the data does live at this location on
disk or get the real location of the data.

The block headers are simply a series of 32 bit chunks, containing the three
fields from a block ref and a location field. The location field will either be
0, signifying the data for this block ref is at this location, or a non-zero
integer, which is the location of a different set of block headers to go to.

When writing a block, if the default location is used for a preexisting block
(so, a hash collision), linear probing is then used to find the next free
location to use.